chore: rename files and move type-guards.ts to another directory
