+++ src/common/utils/type-guards.ts
