--- src/env.d.ts
@@ -1,13 +1,0 @@
-/// <reference types="vite/client" />
-type ImportMetaEnvironment = {
-    readonly VITE_CTP_PROJECT_KEY: string;
-    readonly VITE_CTP_CLIENT_ID: string;
-    readonly VITE_CTP_CLIENT_SECRET: string;
-    readonly VITE_CTP_API_URL: string;
-    readonly VITE_CTP_AUTH_URL: string;
-    readonly VITE_CTP_SCOPES: string;
-};
-
-type ImportMeta = {
-    readonly env: ImportMetaEnvironment;
-};
